refactor events to be handled after other effect features
