+++ js/$wire.js
@@ -1,11 +1,11 @@
-import { dispatch, dispatchSelf, dispatchTo, listen } from '@/features/supportEvents'
+import { removeUpload, upload, uploadMultiple } from './features/supportFileUploads'
+import { dispatch, dispatchSelf, dispatchTo, listen } from '@/events'
 import { generateEntangleFunction } from '@/features/supportEntangle'
 import { closestComponent, findComponent } from '@/store'
 import { requestCommit, requestCall } from '@/request'
 import { WeakBag, dataGet, dataSet } from '@/utils'
-import { on, trigger } from '@/events'
+import { on, trigger } from '@/hooks'
 import Alpine from 'alpinejs'
-import { removeUpload, upload, uploadMultiple } from './features/supportFileUploads'
 
 let properties = {}
 let fallback
