+++ js/directives/wire-model.js
@@ -1,5 +1,5 @@
 import { directive } from '@/directives'
-import { on } from '@/events'
+import { on } from '@/hooks'
 import { handleFileUpload } from '@/features/supportFileUploads'
 import { closestComponent } from '@/store'
 import { dataGet, dataSet } from '@/utils'
