+++ js/directives/wire-transition.js
@@ -1,5 +1,5 @@
 import { directive } from "@/directives"
-import { on } from '@/events'
+import { on } from '@/hooks'
 import Alpine from 'alpinejs'
 
 on('morph.added', ({ el }) => {
