+++ js/features/supportEvents.js
@@ -1,6 +1,5 @@
-import { componentsByName, findComponent } from '@/store'
-import { on as hook, trigger } from '@/events'
-import Alpine from 'alpinejs'
+import { dispatch, dispatchSelf, dispatchTo } from '@/events'
+import { on as hook } from '@/hooks'
 
 hook('effects', (component, effects) => {
     registerListeners(component, effects.listeners || [])
@@ -44,51 +43,4 @@ function dispatchEvents(component, dispatches) {
     })
 }
 
-function dispatchEvent(target, name, params, bubbles = true) {
-    let e = new CustomEvent(name, { bubbles, detail: params })
 
-    e.__livewire = { name, params, receivedBy: [] }
-
-    target.dispatchEvent(e)
-}
-
-export function dispatch(component, name, params) {
-    dispatchEvent(component.el, name, params)
-}
-
-export function dispatchGlobal(name, params) {
-    dispatchEvent(window, name, params)
-}
-
-export function dispatchSelf(component, name, params) {
-    dispatchEvent(component.el, name, params, false)
-}
-
-export function dispatchTo(componentName, name, params) {
-    let targets = componentsByName(componentName)
-
-    targets.forEach(target => {
-        dispatchEvent(target.el, name, params, false)
-    })
-}
-
-export function listen(component, name, callback) {
-    component.el.addEventListener(name, e => {
-        callback(e.detail)
-    })
-}
-
-export function on(eventName, callback) {
-    let handler = (e) => {
-        // Implemented for backwards compatibility...
-        if (! e.__livewire) return
-
-        callback(e.detail)
-    }
-
-    window.addEventListener(eventName, handler)
-
-    return () => {
-        window.removeEventListener(eventName, handler)
-    }
-}
