+++ js/features/supportLaravelEcho.js
@@ -1,5 +1,5 @@
-import { on } from '@/events'
-import { dispatchSelf } from './supportEvents'
+import { on } from '@/hooks'
+import { dispatchSelf } from '@/events'
 
 on('request', ({ options }) => {
     if (window.Echo) {
