+++ js/features/supportScriptsAndAssets.js
@@ -1,4 +1,4 @@
-import { on } from '@/events'
+import { on } from '@/hooks'
 import Alpine from 'alpinejs'
 
 let executedScripts = new WeakMap
@@ -7,7 +7,7 @@ let executedAssets = new Set
 
 on('payload.intercept', async ({ assets }) => {
     if(!assets) return
-    
+
     for (let [key, asset] of Object.entries(assets)) {
         await onlyIfAssetsHaventBeenLoadedAlreadyOnThisPage(key, async () => {
             await addAssetsToHeadTagOfPage(asset)
