+++ js/request/bus.js
@@ -1,4 +1,4 @@
-import { on, trigger } from '@/events'
+import { on, trigger } from '@/hooks'
 import { RequestPool } from './pool'
 import { Commit } from './commit'
